Fix mqttServer String
When trying to connect to a mqtt cloud broker (e.g. cloudmqtt) I need to configure hostname, port, user, password. 
In this case the string of the server is not assembled correctly. The port is missing.
With this fix the port will be appended to the server string and the connection works immediately.